+++ serde_core/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
-name = "serde"
+name = "serde_core"
 version = "1.0.219"
 authors = ["Erick Tryzelaar <erick.tryzelaar@gmail.com>", "David Tolnay <dtolnay@gmail.com>"]
 build = "build.rs"
 categories = ["encoding", "no-std", "no-std::no-alloc"]
-description = "A generic serialization/deserialization framework"
-documentation = "https://docs.rs/serde"
+description = "Core functionalities and abstractions for the Serde serialization/deserialization framework"
+documentation = "https://docs.rs/serde_core"
 edition = "2021"
 homepage = "https://serde.rs"
 keywords = ["serde", "serialization", "no_std"]
@@ -14,18 +14,14 @@ readme = "crates-io.md"
 repository = "https://github.com/serde-rs/serde"
 rust-version = "1.56"
 
-[dependencies]
-serde_core = { version = "=1.0.219", path = "../serde_core", default-features = false }
-serde_derive = { version = "1", optional = true, path = "../serde_derive" }
-
 [dev-dependencies]
-serde_derive = { version = "1", path = "../serde_derive" }
+serde = { version = "1", path = "../serde" }
 
 [package.metadata.playground]
-features = ["derive", "rc"]
+features = ["rc"]
 
 [package.metadata.docs.rs]
-features = ["derive", "rc", "unstable"]
+features = ["rc", "unstable"]
 targets = ["x86_64-unknown-linux-gnu"]
 rustdoc-args = [
     "--generate-link-to-definition",
@@ -34,39 +30,28 @@ rustdoc-args = [
     "--extern-html-root-url=std=https://doc.rust-lang.org",
 ]
 
-# This cfg cannot be enabled, but it still forces Cargo to keep serde_derive's
-# version in lockstep with serde's, even if someone depends on the two crates
-# separately with serde's "derive" feature disabled. Every serde_derive release
-# is compatible with exactly one serde release because the generated code
-# involves nonpublic APIs which are not bound by semver.
-[target.'cfg(any())'.dependencies]
-serde_derive = { version = "=1.0.219", path = "../serde_derive" }
-
 
 ### FEATURES #################################################################
 
 [features]
 default = ["std"]
 
-# Provide derive(Serialize, Deserialize) macros.
-derive = ["serde_derive"]
-
 # Provide impls for common standard library types like Vec<T> and HashMap<K, V>.
 # Requires a dependency on the Rust standard library.
-std = ["serde_core/std"]
+std = []
 
 # Provide impls for types that require unstable functionality. For tracking and
 # discussion of unstable functionality please refer to this issue:
 #
 #    https://github.com/serde-rs/serde/issues/812
-unstable = ["serde_core/unstable"]
+unstable = []
 
 # Provide impls for types in the Rust core allocation and collections library
 # including String, Box<T>, Vec<T>, and Cow<T>. This is a subset of std but may
 # be enabled without depending on all of std.
-alloc = ["serde_core/alloc"]
+alloc = []
 
 # Opt into impls for Rc<T> and Arc<T>. Serializing and deserializing these types
 # does not preserve identity and may result in multiple copies of the same data.
 # Be sure that this is what you want before enabling this feature.
-rc = ["serde_core/rc"]
+rc = []
