+++ serde_core/src/de/size_hint.rs
@@ -1,6 +1,8 @@
+//! Provides helpers for creating size hints for container deserialization.
 #[cfg(any(feature = "std", feature = "alloc"))]
 use crate::lib::*;
 
+/// Extracts the exact size of an iterator if it has a known upper bound and it matches the lower bound.
 pub fn from_bounds<I>(iter: &I) -> Option<usize>
 where
     I: Iterator,
@@ -8,6 +10,7 @@ where
     helper(iter.size_hint())
 }
 
+/// Returns conservative size estimate for a container, clamping the result to a maximum size.
 #[cfg(any(feature = "std", feature = "alloc"))]
 pub fn cautious<Element>(hint: Option<usize>) -> usize {
     const MAX_PREALLOC_BYTES: usize = 1024 * 1024;
