+++ serde_core/src/ser/fmt.rs
