+++ serde_core/src/ser/impossible.rs
